Basic Http server / proxy skeleton completed.
The webserver can accept basic GET requests and read static
content from the disk. From a proxy standpoint, we can "forward"
the basic request line from the original Http Request, but we are
not sending the "complete request" as of yet. The response from
the proxy service is merely writing to the logs, and a new
and independant 404 Reponse is sent back to the original client.